refactor(125): refactor tutorial content structure
