+++ app/src/main/java/org/cescfe/numpairs/feature/tutorial/TutorialContent.kt
@@ -1,0 +1,19 @@
+package org.cescfe.numpairs.feature.tutorial
+
+object TutorialContent {
+    val scenarios: List<TutorialScenario> = listOf(
+        LearnBasicsTutorialContent.scenario,
+        SolvingTipsPracticeContent.scenario
+    )
+
+    val learnBasicsSteps: List<TutorialStep> = LearnBasicsTutorialContent.steps
+    val solvingTipsPracticeSteps: List<TutorialStep> = SolvingTipsPracticeContent.steps
+    val steps: List<TutorialStep> = learnBasicsSteps + solvingTipsPracticeSteps
+
+    fun stepsFor(mode: TutorialMode): List<TutorialStep> = when (mode) {
+        TutorialMode.LEARN_BASICS -> learnBasicsSteps
+        TutorialMode.SOLVING_TIPS_PRACTICE -> solvingTipsPracticeSteps
+    }
+
+    fun scenario(id: TutorialScenarioId): TutorialScenario = scenarios.first { scenario -> scenario.id == id }
+}
